requirements.txt
streamlit>=1.28.0
plotly>=5.15.0
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.18
scipy>=1.11.0
python-dotenv>=1.0.0
alpaca-trade-api>=3.0.0
polygon-api-client>=1.12.0
requests>=2.31.0
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0

